Hotfix: Admin Panel - Admin can change userId in services. - If the category in the usage admin user can't delete it from DB.
