+++ app/Http/Controllers/API/v1/Admin/AdminCityController.php
@@ -1,37 +1,34 @@
 <?php
 
-namespace App\Http\Controllers\API\v1;
+namespace App\Http\Controllers\API\v1\Admin;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Service\ServiceDestroyRequest;
-use App\Http\Requests\Service\ServiceIndexRequest;
-use App\Http\Requests\Service\ServiceShowRequest;
-use App\Http\Requests\Service\ServiceStoreRequest;
-use App\Http\Requests\Service\ServiceUpdateRequest;
-use App\Http\Resources\Service\ServiceResource;
-use App\Repositories\Services\ServiceStoreDTO;
-use App\Repositories\Services\ServiceUpdateDTO;
-use App\Services\Service\ServiceService;
+use App\Http\Requests\City\CityDestroyRequest;
+use App\Http\Requests\City\CityIndexRequest;
+use App\Http\Requests\City\CityShowRequest;
+use App\Http\Requests\City\CityStoreRequest;
+use App\Http\Requests\City\CityUpdateRequest;
+use App\Http\Resources\City\CityResource;
+use App\Repositories\Cities\CityStoreDTO;
+use App\Repositories\Cities\CityUpdateDTO;
+use App\Services\City\CityService;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Response;
 use OpenApi\Attributes as OA;
 
-class ServiceController extends Controller
+class AdminCityController extends Controller
 {
-    /**
-     * @param ServiceService $service
-     */
     public function __construct(
-        protected ServiceService $service,
+        protected CityService $cityService,
     ) {
     }
 
     /**
-     * @param ServiceIndexRequest $request
+     * @param CityIndexRequest $request
      * @return JsonResponse
      */
     #[OA\Get(
-        path: '/v1/admin/services',
+        path: '/v1/admin/cities',
         security: [['bearerAuth' => []]],
         tags: ['Admin Panel'],
         parameters: [
@@ -48,12 +45,12 @@ class ServiceController extends Controller
         responses: [
             new OA\Response(
                 response: 200,
-                description: 'Show all services',
+                description: 'Show all cities',
                 content: new OA\JsonContent(
                     properties: [
                         new OA\Property(
                             property: 'data',
-                            ref: '#/components/schemas/Service',
+                            ref: '#/components/schemas/City',
                         ),
                     ],
                 ),
@@ -67,84 +64,61 @@ class ServiceController extends Controller
             ),
         ],
     )]
-    public function index(ServiceIndexRequest $request): JsonResponse
+    public function index(CityIndexRequest $request): JsonResponse
     {
         $validated = $request->validated();
-        $service = $this->service->getPrivateServices($validated['lastId']);
-        $resource = ServiceResource::collection($service);
+        $service = $this->cityService->getAllCitiesByLastId($validated['lastId']);
+        $resource = CityResource::collection($service);
 
         return $resource->response()->setStatusCode(200);
     }
 
     /**
-     * @param ServiceStoreRequest $request
+     * @param CityStoreRequest $request
      * @return JsonResponse
      */
     #[OA\Post(
-        path: '/v1/admin/services',
+        path: '/v1/admin/cities',
         security: [['bearerAuth' => []]],
         tags: ['Admin Panel'],
         parameters: [
             new OA\Parameter(
-                name: 'categoryId',
-                description: 'Only existing Category ID',
+                name: 'countryId',
+                description: 'Country ID',
                 in: 'query',
                 required: true,
                 schema: new OA\Schema(
                     type: 'integer',
-                    minimum: 1,
                 ),
             ),
             new OA\Parameter(
-                name: 'title',
-                description: 'max: 255',
+                name: 'parentId',
+                description: 'Parent Category ID',
                 in: 'query',
-                required: true,
-                schema: new OA\Schema(
-                    type: 'string',
-                    maxLength: 255,
-                ),
-            ),
-            new OA\Parameter(
-                name: 'description',
-                description: 'max: 500',
-                in: 'query',
-                required: true,
-                schema: new OA\Schema(
-                    type: 'string',
-                    maxLength: 500,
-                ),
-            ),
-            new OA\Parameter(
-                name: 'price',
-                description: 'regex in format: 999.99',
-                in: 'query',
-                required: true,
-                schema: new OA\Schema(
-                    type: 'number',
-                    format: 'double',
-                ),
-            ),
-            new OA\Parameter(
-                name: 'cityId',
-                description: 'Only existing City ID',
-                in: 'query',
-                required: true,
                 schema: new OA\Schema(
                     type: 'integer',
-                    minimum: 1,
+                    nullable: true,
                 ),
+            ),
+            new OA\Parameter(
+               name: 'name',
+                description: 'Name of the city',
+               in: 'query',
+               required: true,
+               schema: new OA\Schema(
+                   type: 'string',
+               ),
             ),
         ],
         responses: [
             new OA\Response(
                 response: 201,
-                description: 'Create new service',
+                description: 'Created city',
                 content: new OA\JsonContent(
                     properties: [
                         new OA\Property(
                             property: 'data',
-                            ref: '#/components/schemas/Service',
+                            ref: '#/components/schemas/City',
                         ),
                     ],
                 ),
@@ -158,44 +132,44 @@ class ServiceController extends Controller
             ),
         ],
     )]
-    public function store(ServiceStoreRequest $request): JsonResponse
+    public function store(CityStoreRequest $request): JsonResponse
     {
         $validated = $request->validated();
-        $DTO = new ServiceStoreDTO(...$validated);
-        $service = $this->service->insertAndGetService($DTO);
-        $resource = new ServiceResource($service);
+        $DTO = new CityStoreDTO(...$validated);
+        $service = $this->cityService->insertAndGetCity($DTO);
+        $resource = new CityResource($service);
 
         return $resource->response()->setStatusCode(201);
     }
 
     /**
-     * @param ServiceShowRequest $request
+     * @param CityShowRequest $request
      * @return JsonResponse
      */
     #[OA\Get(
-        path: '/v1/admin/services/{id}',
+        path: '/v1/admin/cities/{id}',
         security: [['bearerAuth' => []]],
         tags: ['Admin Panel'],
         parameters: [
             new OA\Parameter(
                 name: 'id',
+                description: 'City ID',
                 in: 'path',
                 required: true,
                 schema: new OA\Schema(
                     type: 'integer',
-                    minimum: 1,
                 ),
             ),
         ],
         responses: [
             new OA\Response(
                 response: 200,
-                description: 'Show the service',
+                description: 'Show information about the city',
                 content: new OA\JsonContent(
                     properties: [
                         new OA\Property(
                             property: 'data',
-                            ref: '#/components/schemas/Service',
+                            ref: '#/components/schemas/City',
                         ),
                     ],
                 ),
@@ -209,93 +183,70 @@ class ServiceController extends Controller
             ),
         ],
     )]
-    public function show(ServiceShowRequest $request): JsonResponse
+    public function show(CityShowRequest $request): JsonResponse
     {
         $validated = $request->validated();
-        $service = $this->service->getById($validated['id']);
-        $resource = new ServiceResource($service);
+        $service = $this->cityService->getById($validated['id']);
+        $resource = new CityResource($service);
 
         return $resource->response()->setStatusCode(200);
     }
 
     /**
-     * @param ServiceUpdateRequest $request
+     * @param CityUpdateRequest $request
      * @return JsonResponse
      */
     #[OA\Patch(
-        path: '/v1/admin/services/{id}',
+        path: '/v1/admin/cities/{id}',
         security: [['bearerAuth' => []]],
         tags: ['Admin Panel'],
         parameters: [
             new OA\Parameter(
                 name: 'id',
+                description: 'City ID',
                 in: 'path',
                 required: true,
                 schema: new OA\Schema(
                     type: 'integer',
-                    minimum: 1,
                 ),
             ),
             new OA\Parameter(
-                name: 'categoryId',
-                description: 'Only existing Category ID',
+                name: 'countryId',
+                description: 'Country ID',
                 in: 'query',
                 required: true,
                 schema: new OA\Schema(
                     type: 'integer',
-                    minimum: 1,
                 ),
             ),
             new OA\Parameter(
-                name: 'title',
-                description: 'max: 255',
+                name: 'parentId',
+                description: 'Parent Category ID',
                 in: 'query',
-                required: true,
-                schema: new OA\Schema(
-                    type: 'string',
-                    maxLength: 255,
-                ),
-            ),
-            new OA\Parameter(
-                name: 'description',
-                description: 'max: 500',
-                in: 'query',
-                required: true,
-                schema: new OA\Schema(
-                    type: 'string',
-                    maxLength: 500,
-                ),
-            ),
-            new OA\Parameter(
-                name: 'price',
-                description: 'regex in format: 999.99',
-                in: 'query',
-                required: true,
-                schema: new OA\Schema(
-                    type: 'number',
-                    format: 'double',
-                ),
-            ),
-            new OA\Parameter(
-                name: 'cityId',
-                description: 'Only existing City ID',
-                in: 'query',
-                required: true,
                 schema: new OA\Schema(
                     type: 'integer',
-                    minimum: 1,
+                    nullable: true,
+                ),
+            ),
+            new OA\Parameter(
+                name: 'name',
+                description: 'Name of the city',
+                in: 'query',
+                required: true,
+                schema: new OA\Schema(
+                    type: 'string',
                 ),
             ),
         ],
         responses: [
             new OA\Response(
                 response: 200,
-                description: 'Show updated service',
+                description: 'Show information about updated city',
                 content: new OA\JsonContent(
                     properties: [
                         new OA\Property(
                             property: 'data',
-                            ref: '#/components/schemas/Service',
+                            ref: '#/components/schemas/City',
                         ),
                     ],
                 ),
@@ -309,46 +260,53 @@ class ServiceController extends Controller
             ),
         ],
     )]
-    public function update(ServiceUpdateRequest $request): JsonResponse
+    public function update(CityUpdateRequest $request): JsonResponse
     {
         $validated = $request->validated();
-        $DTO = new ServiceUpdateDTO(...$validated);
-        $service = $this->service->updateAndGetById($DTO);
-        $resource = new ServiceResource($service);
+        $DTO = new CityUpdateDTO(...$validated);
+        $service = $this->cityService->update($DTO);
+        $resource = new CityResource($service);
 
         return $resource->response()->setStatusCode(200);
     }
 
     /**
-     * @param ServiceDestroyRequest $request
+     * @param CityDestroyRequest $request
      * @return Response
      */
     #[OA\Delete(
-        path: '/v1/admin/services/{id}',
+        path: '/v1/admin/cities/{id}',
         security: [['bearerAuth' => []]],
         tags: ['Admin Panel'],
         parameters: [
             new OA\Parameter(
                 name: 'id',
+                description: 'City ID',
                 in: 'path',
                 required: true,
                 schema: new OA\Schema(
                     type: 'integer',
-                    minimum: 1,
                 ),
             ),
         ],
         responses: [
             new OA\Response(
                 response: 204,
-                description: 'Delete the service',
+                description: 'The city has been deleted',
+            ),
+            new OA\Response(
+                response: 422,
+                description: 'Validation errors',
+                content: new OA\JsonContent(
+                    ref: '#/components/schemas/ValidationErrors'
+                )
             ),
         ],
     )]
-    public function destroy(ServiceDestroyRequest $request): Response
+    public function destroy(CityDestroyRequest $request): Response
     {
         $validated = $request->validated();
-        $this->service->deleteById($validated['id']);
+        $this->cityService->delete($validated['id']);
 
         return response()->noContent();
     }
