+++ app/Http/Controllers/API/v1/Authentication/EmailVerificationController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\API\v1;
+namespace App\Http\Controllers\API\v1\Authentication;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Authentication\EmailVerificationRequest;
