+++ requests/v1/Categories/categories.destroy.http
@@ -1,4 +1,4 @@
-DELETE http://{{host}}/api/v1/categories/{{id}}
+DELETE http://{{host}}/api/v1/admin/categories/{{id}}
 Accept: application/json
 Content-Type: application/json
 Authorization: {{token}}
