+++ requests/v1/Categories/categories.show.http
@@ -1,4 +1,4 @@
-GET http://{{host}}/api/v1/categories/{{id}}
+GET http://{{host}}/api/v1/admin/categories/{{id}}
 Accept: application/json
 Content-Type: application/json
 Authorization: {{token}}
