+++ requests/v1/Categories/categories.store.http
@@ -1,4 +1,4 @@
-POST http://{{host}}/api/v1/categories
+POST http://{{host}}/api/v1/admin/categories
 Accept: application/json
 Content-Type: application/json
 Authorization: {{token}}
