+++ requests/v1/Cities/cities.destroy.http
@@ -1,4 +1,4 @@
-DELETE http://{{host}}/api/v1/cities/{{id}}
+DELETE http://{{host}}/api/v1/admin/cities/{{id}}
 Accept: application/json
 Content-Type: application/json
 Authorization: {{token}}
