+++ requests/v1/Cities/cities.index.http
@@ -1,4 +1,4 @@
-GET http://{{host}}/api/v1/cities
+GET http://{{host}}/api/v1/admin/cities
 Accept: application/json
 Content-Type: application/json
 Authorization: {{token}}
