+++ requests/v1/Cities/cities.show.http
@@ -1,4 +1,4 @@
-GET http://{{host}}/api/v1/cities/{{id}}
+GET http://{{host}}/api/v1/admin/cities/{{id}}
 Accept: application/json
 Content-Type: application/json
 Authorization: {{token}}
