+++ routes/api.php
@@ -1,12 +1,12 @@
 <?php
 
-use App\Http\Controllers\API\v1\AuthenticationController;
-use App\Http\Controllers\API\v1\CategoryController;
-use App\Http\Controllers\API\v1\CityController;
-use App\Http\Controllers\API\v1\EmailVerificationController;
+use App\Http\Controllers\API\v1\Admin\AdminCategoryController;
+use App\Http\Controllers\API\v1\Admin\AdminCityController;
+use App\Http\Controllers\API\v1\Admin\AdminServiceController;
+use App\Http\Controllers\API\v1\Authentication\AuthenticationController;
+use App\Http\Controllers\API\v1\Authentication\EmailVerificationController;
+use App\Http\Controllers\API\v1\Authentication\PasswordController;
 use App\Http\Controllers\API\v1\HomePageController;
-use App\Http\Controllers\API\v1\PasswordController;
-use App\Http\Controllers\API\v1\ServiceController;
 use App\Http\Middleware\API\GuestMiddleware;
 use App\Services\Swagger\SwaggerService;
 use Illuminate\Support\Facades\Route;
@@ -46,9 +46,9 @@ Route::group(['prefix' => 'v1'], function () {
         });
 
         Route::group(['prefix' => 'admin', 'middleware' => ['isAuthUserAdmin']], function () {
-            Route::apiResource('cities', CityController::class);
-            Route::apiResource('categories', CategoryController::class);
-            Route::apiResource('services', ServiceController::class);
+            Route::apiResource('cities', AdminCityController::class);
+            Route::apiResource('categories', AdminCategoryController::class);
+            Route::apiResource('services', AdminServiceController::class);
         });
     });
